Update Shift Zero To The End
