+++ kotlin/src/test/kotlin/common/twopointers/shiftzero/ShiftZeroTest.kt
@@ -5,20 +5,12 @@ import kotlin.test.assertContentEquals
 
 class ShiftZeroTest {
 
-    private val question = ShiftZero()
+    private val question = ShiftZeroToTheEnd()
 
     @Test
-    fun `Test solution 1`() {
+    fun `Test solution`() {
         val input = intArrayOf(0, 5, 4, 0, 4, 5, 1, 2, 7, 0)
         question.solution(input)
         assertContentEquals(intArrayOf(5, 4, 4, 5, 1, 2, 7, 0, 0, 0), input)
     }
-
-    @Test
-    fun `Test solution 2`() {
-        val input = intArrayOf(0, 5, 4, 0, 4, 5, 1, 2, 7, 0)
-        question.solution2(input)
-        assertContentEquals(intArrayOf(5, 4, 4, 5, 1, 2, 7, 0, 0, 0), input)
-    }
-
 }
